Move `special-waves.ts` to `src/constants/`
