+++ src/constants/special-waves.ts
