+++ src/phases/victory-phase.ts
@@ -4,7 +4,7 @@ import type { GameOverPhase } from "#app/phases/game-over-phase";
 /* eslint-enable @typescript-eslint/no-unused-vars */
 // -- end tsdoc imports--
 
-import { EVIL_BOSS_2_WAVE } from "#app/data/special-waves";
+import { EVIL_BOSS_2_WAVE } from "#app/constants/special-waves";
 import { globalScene } from "#app/global-scene";
 import type { CustomModifierSettings } from "#app/modifier/modifier-type";
 import { modifierTypes } from "#app/modifier/modifier-types";
